requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.1.0
email-validator>=1.1.3,<1.2.0
bcrypt>=3.2.0,<3.3.0
alembic>=1.7.0,<1.8.0
starlette~=0.14.2
colorlog~=6.9.0
slowapi~=0.1.9